+++ generator/src/main.rs
@@ -1,3 +1,4 @@
+use crate::plt::{CreatePltGenerator, PltOperationGenerator};
 use anyhow::Context;
 use clap::{Parser, Subcommand};
 use concordium_rust_sdk::{endpoints::Endpoint, types::WalletAccount, v2};
@@ -5,9 +6,10 @@ use generator::{
     generate_transactions, CcdGenerator, CommonArgs, MintCis2Generator,
     RegisterCredentialsGenerator, RegisterDataGenerator, TransferCis2Generator, WccdGenerator,
 };
-use std::path::PathBuf;
+use std::{path::PathBuf, time::Duration};
 
 mod generator;
+mod plt;
 
 #[derive(clap::Parser, Debug)]
 #[clap(author, version, about)]
@@ -18,21 +20,26 @@ struct App {
         help = "GRPC interface of the node.",
         default_value = "http://localhost:20000"
     )]
-    endpoint: Endpoint,
+    endpoint:      Endpoint,
     #[clap(long = "sender", help = "Path to file containing sender keys.")]
-    account:  PathBuf,
+    account:       PathBuf,
     #[clap(
         long = "tps",
         help = "Transactions to send per second.",
         default_value = "1"
     )]
-    tps:      u16,
+    tps:           u16,
+    #[clap(
+        long = "send-interval",
+        help = "Time between transactions in microseconds. If specified, overrides 'tps' argument."
+    )]
+    send_interval: Option<u64>,
     #[clap(
         long = "expiry",
         help = "Expiry of transactions in seconds.",
         default_value = "7200"
     )]
-    expiry:   u32,
+    expiry:        u32,
 
     #[command(subcommand)]
     command: Command,
@@ -54,6 +61,10 @@ enum Command {
     RegisterCredentials,
     /// Register data
     RegisterData(generator::RegisterDataArgs),
+    /// PLT update operations
+    Plt(plt::PltOperationArgs),
+    /// Create PLT update instructions
+    CreatePlt(plt::CreatePltArgs),
 }
 
 #[tokio::main(flavor = "multi_thread")]
@@ -83,34 +94,50 @@ async fn main() -> anyhow::Result<()> {
         expiry: app.expiry,
     };
 
+    let txn_interval = if let Some(send_interval) = app.send_interval {
+        Duration::from_micros(send_interval)
+    } else {
+        Duration::from_micros(1_000_000 / u64::from(app.tps))
+    };
+
     match app.command {
         Command::Ccd(ccd_args) => {
             let generator = CcdGenerator::instantiate(client.clone(), args, ccd_args).await?;
-            generate_transactions(client, generator, app.tps).await
+            generate_transactions(client, generator, txn_interval).await
         }
         Command::MintNfts => {
             let generator = MintCis2Generator::instantiate(client.clone(), args).await?;
-            generate_transactions(client, generator, app.tps).await
+            generate_transactions(client, generator, txn_interval).await
         }
         Command::TransferCis2(transfer_cis2_args) => {
             let generator =
                 TransferCis2Generator::instantiate(client.clone(), args, transfer_cis2_args)
                     .await?;
-            generate_transactions(client, generator, app.tps).await
+            generate_transactions(client, generator, txn_interval).await
         }
         Command::Wccd => {
             let generator = WccdGenerator::instantiate(client.clone(), args).await?;
-            generate_transactions(client, generator, app.tps).await
+            generate_transactions(client, generator, txn_interval).await
         }
         Command::RegisterCredentials => {
             let generator = RegisterCredentialsGenerator::instantiate(client.clone(), args).await?;
-            generate_transactions(client, generator, app.tps).await
+            generate_transactions(client, generator, txn_interval).await
         }
         Command::RegisterData(register_data_args) => {
             let generator =
                 RegisterDataGenerator::instantiate(client.clone(), args, register_data_args)
                     .await?;
-            generate_transactions(client, generator, app.tps).await
+            generate_transactions(client, generator, txn_interval).await
+        }
+        Command::Plt(plt_args) => {
+            let generator =
+                PltOperationGenerator::instantiate(client.clone(), args, plt_args).await?;
+            generate_transactions(client, generator, txn_interval).await
+        }
+        Command::CreatePlt(create_plt_args) => {
+            let generator =
+                CreatePltGenerator::instantiate(client.clone(), args, create_plt_args).await?;
+            generate_transactions(client, generator, txn_interval).await
         }
     }
 }